t/crispr.t added code to use TestMethods methods for a test genome file, bwa index and annotation file.
t/lib/TestMethods.pm
New File: TestMethods - module for checking for a test genome file, bwa index
and annotation file.
Method: check_for_test_genome - Checks for genome file.
If it doesn't exist, it connects to Ensembl and creates
a fasta file containing the sequence for zebrafish chromosome 3.
It then creates a bwa index for the fasta file if it doesn't exist.
Method: check_for_annotation - Checks for a test annotation file.
If it doesn't exist, it connects to Ensembl and creates
a gff file containing exon and intron data for fish chr3.